+++ trees/decision/__init__.py
@@ -19,44 +19,46 @@
 # limitations under the License.
 #
 
-from trees import NodeProcessor
+from trees import NodeProcessor, Node
 
-class GraphvizProcessor(NodeProcessor):
+class DecisionProcessor(NodeProcessor):
+
+    def __init__(self, X, Y, stop_fun, split_fun):
+        self.X = X
+        self.Y = Y
+        self.stop_fun = stop_fun
+        self.split_fun = split_fun
 
     def root_init(self, root):
-        self.nbr = 0
-        self.graph = "digraph Tree {\nnode [shape=box] ;\n"
-        self.father_id = []
+        setattr(root, "X", self.X)
+        setattr(root, "Y", self.Y)
 
     def root_final(self, root):
-        self.graph += "}"
-
-    def _fill_node(self, node):
-        self.nbr += 1
-        nid = 'N' + str(self.nbr)
-        self.graph += nid + '[label="' + node.label + '"];\n'
-        return nid
+        pass
 
     def child_pre_process(self, node, key):
-        nid = self._fill_node(node.children[key])
-        self.graph += self.father_id[-1] + ' -> ' + nid + ' [label="' + key + '"];\n'
-        self.father_id.append(nid)
         return False
 
     def child_post_process(self, node, child_key):
-        self.father_id.pop()
         return False
 
     def pre_process(self, node):
-        if len(self.father_id) == 0:
-            self.father_id.append(self._fill_node(node))
+        test, test_res, attr = self.stop_fun(node.X, node.Y)
+        if test:
+            for val in attr:
+                setattr(node, val, attr[val])
+            return []
+        sets = self.split_fun(node.X, node.Y, test_res)
+        for key in sets:
+            the_set = sets[key]
+            child = Node(value=the_set.value, label=the_set.label)
+            setattr(child, "X", the_set.X)
+            setattr(child, "Y", the_set.Y)
+            node.children[key] = child
         return node.children.keys()
 
     def post_process(self, node):
-        #self.father_id.pop()
         pass
 
     def result(self):
-        return self.graph
-
-graphviz_processor = GraphvizProcessor()
+        pass